requirements.txt
Django==3.2
djangorestframework==3.12.4
Pyrebase==3.0.27
pylint==2.7.4
asgiref==3.3.4
psycopg2==2.8.6
Pillow==8.2.0
whitenoise==5.2.0
cryptography==2.8
gunicorn==20.1.0
dj-database-url==0.5.0
requests==2.11.1